add endpoint with tests for fetching activity by id

+++ Sources/StravaApi/Models/MetaActivity.swift
@@ -5,5 +5,5 @@ import Foundation
 /// source: https://developers.strava.com/docs/reference/#api-models-MetaActivity
 public struct MetaActivity: Decodable, Equatable {
     ///The unique identifier of the activity
-    public let id: Double?
+    public let id: Int?
 }
