+++ Sources/StravaApi/Models/PhotosSummary.swift
@@ -4,7 +4,7 @@ import Foundation
 /// This is a swift representation of current PhotosSummary
 /// source: https://developers.strava.com/docs/reference/#api-models-PhotosSummary
 public struct PhotosSummary: Decodable, Equatable {
-    public let id: Double?
+    public let id: Int?
     public let source: Int?
     public let unique_id: String?
     public let urls: String?
